Renamed parallel_raste_processor.* files for consistency.
