+++ src/raster_parallel_processor.h
